Added new DIY tutorial page
Added the new DIY Tutorial page. With a very basic text area and
mdxeditor for editing the prompt.
Switched to mobx from zustand as I thought that zustand might lead to
performance issues and I want to keep the architecture as close to
wordcraft as possible.
Also, included the material design components and set the color pallette
for the app. So that I dont have to worry too much about basic
components and color pallette and I can just focus on developing stuff.