Converting the project to lit
Just going ahead with the dependencies from wordcraft. Will Modify names
and placeholder for the diy-tutorial-mate. But otherwise copy everything
else from wordcraft and just focus on working on differences instead of
starting everything from scratch. I should not waste time trying to
re-engineer code from it. It would be faster to learn the new libraries
and adapt it to do what needs to be done.