Will continue ahead with Lit+Mobx+vite
I got hmr for the vite plugin. This allows me to develop stuff without
the state resetting all the time. This was a major issue in the existing
wordcraft project along with the fact that they had a ton of issues in
terms of outdated packages whose documentation was inaccessible or hard
to figure out. With this it should be allright to continue ahead and
develop I also got a router for lit so this should allow me to handle
the page handling and stuff. I should focus on making the project and
should not focus on build issues. I now know how to do hmr for lit
elements and will use that develop things properly.