Add Router Service and enabled service creation
Added a router service to keep track of pages and help make that
routing simpler and consolidate all of that information into a single
place. Also, this allows me to create a service based architecture where
I can add services and that enables me to work ahead from there. I can
continue ahead with adding session and creating the architecture to
allow for different AI models to be added.