Connected the backend & adds outline prompt
Worked on adding the outline prompt and connecting the frontend to the
backend. Created the context to help allow me to add some examples.
Also, started added model service,initialization service, local storage
service and context service. Allowing me to create a service based
architecture and splitting the UI to just be dumb components and
shifting all the business logic into the models and services.