Started working on the main editor page
Started working on the main editor page. Integrated the markdown plugin
into the text editor service. It works well. I was running into a weird
issue with an empty editor state. But clearing the local storage key
resolved the issue. I initially thought it was an issue with the
markdown plugin but It was just issues with my code.
Currently setup the editor page. I now have to work on the sidebar
components and stuff. I will try and integrate as much wordcraft code as
required to try and complete the baseline so that I can get to the
features that I want to work on.