Started working on the cursor service
Trying to get the data from the editor out  of lexical so that I can
use this for the model service and use them in prompts. Mostly, trying
to translate the methods in the original wordcraft one into things from
lexical. Using GPT-4 for ideation helps a lot to get over smaller
anxiety hurdles.