Sentence detection and highlight current sentence
Got the hang for lexical editor. Also, was able to run the sentence
parser and got a hang for lexical now I need to start working on the
feature to get stuff from the cursor service and text service and then
give it to models for their prompts and stuff. I might try to simplify
the document structure but next i need to work on Operation service
which will be responsible for figuring out available operations and also
the chat service it might be easier with the chat service to do things.
But I should finally start working on the sidebar.

Figured out why my sentence detection stuff wasnt working or going into
weird infinite loops apparently I needed toggle unmergeable and it
resolved things for me.