Finished the chat tab
Finished working on the chat tab. Fixed issues in the backend in the
chat endpoint. But I am happy that I finished the chat stuff. I think
progress will be quick if I dont spend time on needless time sinks like
trying to highlight the current sentence I think I will eventually
figure it out.
Chat tab works and sends the current DIY tutorial and description and
also allows the user to chat with the assitant. I was even able to get
it to review stuff and answer quesitons related to it.
Also, worked on the sidebar where you have the controls chat and other
stuff. The main editor page is starting to take shape.