Fixed cursorOffset & moved cursorUpdate to cmd
Fixed the cursor offset since for some reason it was carrying an earlier
value. added a parentKey to the current node and I am using that to set
the cursorOffset. Also, shifted the cursor update to on selection
change. This makes sense as in the read loop it wound up in an infinite
loop if I tried to update anything there. Here I can safely make
changes. Almost there to a solution for the current sentence highlight.
Just need to fix it when creating a new node.
Got it to a pretty good state. Where on new sentences as well it clears
the style.