Added parentOffset calculation to current node.
Turns out the reason selection jumps is because I am not getting the
right selection value after the patch selection. if only I could the get
the value for it and I just need to appropriately set the cursor to the
start or end of the node depending on how they came to it. It is because
it refers to the parent offset that there is an issue there.
Also, fixed a bunch of issues where the cursor offset was lagging
behind. I should really work on the continuation propmpt instead  of
this.