Adds OperationService and DIY operation sites
Added some additional checks in cursorService to track if the cursor is
in introduction, step or conclusion. Also, created a new method in
OperationService to get the place of the cursor in the document and not
just in the sentence or node.
Added DIY specific enums to the OperationSite in types. Also, added
OperationData interface from wordcraft. Will definitely have to simplify
some of the logic from wordcraft its just too overcomplicated for
something that should be simple. Also, I would rather the Operations
themselves directly interact and get the data from the controls and
other services. It is just a bit too much and I should try to simplify
it for my sake so that I can focus on just the prompts and the UI.