Worked on Operations, Document Storage and init
I worked on the operations service and got it to a state with choice
operations I still need to add the review operation. For now I have
created a continuation operation and I still need to check whether it
works now.

I also finished the part where I store and retrieve the documents and
reset the texteditor as well. This was all done to make sure I can
cleanly reboot and work with the next participant. I finished the todo
methods left in the local_storage service. And created and finished the
init service.

Well I am mostly taking stuff directly from wordcraft and adjusting it
to my usecase. I am hopeful to finish the features since the only things
left now are the operations themselves.

Since I ran out of quota. Marcel sent my the api key for the i10 api
key.