Registered Continue Operation and solved errors
Registered the continue operation and fixed the mobx error from when I
tried to directly use wordcraft from before turns out I had to change an
observable from before to an observable.ref so that the computed
function does not panic because it only has proxies and cant call
functions on the classes registered in it. But, I fixed that issue.

Also, fixed the jittering that happens when a new paragraph is created
and a user starts typing in it. I initially wanted it so that it
highlights the current sentence but I realised when the cursor is at the
leading edge of a sentence it actually does not detect the sentence it
is on and thus highlighting the current sentence when the user is
actively typing is counterintuitive and does not work. So I removed the
highlight_sentence call from the textContentListener and removed that
event listener altogether. I also cleaned up and removed the commented
code around the cursor event listeners.

Now, I will work on the components for the operations and then fix up
all the operations for the diymate app. This should take about a week.