Finished the components for choice Operations
Created the UI components required for the various kinds of the
operations. I still have to make a review step. So I will still have to
do that but for now I will focus on the choice operations and finish
them up. Also, copied all the different kinds of controls from
wordcraft. Customised the styling for the diymate application.

Since with the review operations I will probably have to create a
service which stores the data.

Fixed the configuration issues with lit and now I can finally
auto-import the custom web components. Also, set the md-icon-font in the
global css file.

With this the I can finally start to implement the choice operations in
the model dialog.