Converted pre and post text to markdown
Isolated the lexical configuration to a seperate file for lexical.
Fixed mobx Issues in operationsService. Mainly worked on getting the pre
text and post text are in markdown. Also, created new observables for
the cursor with markdown text for the selection and the keys when
selecting multiple element nodes this might be useful to change multiple
elements together for example when inserting a step entirely on its own.
But more importantly this is useful for the getting full text before and
after the cursor to the operations. I can finally test an operation will
have to create new nodes maybe to test. For example loading node etc.