Continuation Operation works but cant insert text
I created the continue prompt handler for the openAI model. It works, I
had to debug an issue with the decorator used in the operation abstract
class.  Turns out I had written a constructor. There was a slight issue
with the UI not updating as the currentStep value changed but creating
the observable in the contructor of the operation abstract class and
annotating currentStep as an observable.ref fixed the issue properly.

I had to remove the can_star stuff in the choicesComponent. Implemented
the text insertion methods in text editor but it is super tedious to do
anything with lexical especially selection operations. I will try and
create a branch where I transition to quill instead. Since it feels a
hell of a lot easier and the api is far more easily serializable. In
general it feels like things would be a lot easier in quill.  Well I
know I said that I locked the libaries but at this point lexical is just
super painful every small thing seems like a hassle to do.

I also changed the name of the selectedText to selectedPlainText and
selectedNodesMarkdownText to selectedMdText.  Modified the operationData
inteface to allow for various different values it will get simplified
once I transition to quill since it is just seems to be so much easier
to use with a wider support base.

Made choices class more compliant to mobx since I was getting errors
when it tried to change index values.

Created the loadingNode and ChoiceNode.