Integrated tiptap editor
It works out of the box without any major issues or having to patch
things. It is also quite easy to programatically set the selection and
things like that.
It is overall a well made editor api and framework. Definitely better
than lexical. Though trying to customise certain things might take some
time. But I am confident it will be far easier than lexical in trying to
get it to do what I want.
I have converted the outline-editor and diymate-editor to tiptap. I also
realised the reason why there were ',' when inserting the html markup
was because in template literals maps of strings are created as an array
and are not joined. so that was what was resulting in my problems. I
definitely need to pick up the pace tomorrow and finish this up so that
I can concentrate on the things that matter like the prompts and other
things.

I am so glad it just works without any issues in the shadowDOM. I am
really suprised by that.