Tiptap fully integrated into services
I have successfully fully integrated tip tap into cursorService and
sentence service. This has been pretty painless. The only issue I had
with regards to it was trying to get the list items to highlight and
track in sentences. This had to do with a peculiar fact about how lists
were implemented. I can finally start concentrating on the features from
tomorrow.

I had to make very minor changes to sentence services and it worked
pretty well had to add a start offset for paragraphData. I still ended
up requiring a custom structure for it. But otherwise it was pretty
fine. Now I need to focus on loading node, choice node and selection
text. Hopefully it will be easy.

I was able to reduce the size of my code though it was pretty neat.