Choice insertion now works properly
Finally, choice insertion works properly as intended. It was way easier
with tiptap I did not have to worry about selection and trying to manage
it.

I also cleaned up and removed all package associations to lexical. I
think I think tomorrow I will start working on the nodes and styling for
them. I will also work on the conclusion generation for the DIY and
replacing the title. It will be a lot easier given that text insertion
works properly and I was able to do it without tearing my hair up.