Created Choice and Loading Nodes
Created Choice and loading nodes. Imported showdown to convert to and
from markdown for the given text. Now I need to work on adding the
different operations.

Converted the sentence service to work without affecting the user
selection. Which is nice. Since it was resulting in weird selection
issues in lists.

Also, added a parser that parses html and converts it to nodes that can
be inserted into content. Since I had to figure out a way to show the
choices without it showing plain html straight up. Inserting html
directly seems to fail if there is more than one node in it. So I have
to parse them into prose mirror nodes.

Adding the operation service any where to restrict the other services so
that they only run a few times here and there messes up a lot of
things. But hopefully since now the current sentence highlights are not
setting the text selection we should be fine.

Isolated the configuration stuff for tiptap into @lib/tiptap.

Removed patch selection from text_editor_service since it is no longer
required.

Also, using custom mark now instead.