Converted all text observables to markdown
Wrote down methods to convert a given piece of the document to markdown
or html based on what we want. It seems to work pretty easily and
effectively.

Essentially I now have methods to convert a selection to html,text or
markdown. I also made it so that it filters out the highlight text
though if I need it I can just retain it. Or get it in a specific
manner. It will be easy though with these methods they will make life a
lot easier.

Now, I can fully focus on the operations tomorrow and get things done
there. I should have switched to tiptap far earlier. Would have saved me
a lot of headache.

Also, changed the mark to a highlight-mark. Made it so that it only
removes that specific mark from sentence and not all marks.