Added next sentence operation, changed DIY format
Added a next sentence operation realised that I will have to use it
differently compared to how the folks in wordcraft were using it. After
some frustration and anxiety changed where it was activated and now it
works well and as intended. Also, had to create different inline nodes
for inline operations. This might be useful when doing replacements
instead.

Also, changed the DIY format a bit to allow for easier isolation of
section. Will have to add a before to the heading to represent # as is
common in markdown. The issue might be trying to get it to look the same
as the heading itself.

Also, because of those changegs to the structure. Had to update cursor
service conditions a bit. Overall it seems easier with fixed sections
instead of free floating sections that might allow for arbitratry stuff
to happen.

I should try to create a document schema that enforces the heading
conditions and automatically changes the attributes to the proper one so
that I dont allow for title headings in the middle of nowhere.