Added elaborate,freeform and replace operations
Added these new operations and prompts. Also, fixed the styling issue in
choices menu. Finally it feels like I am working on the real meat of my
thesis. It has been a long and tiring journey to reach this position.
But, I feel confident that I will finish up with my prototype I should
set things up for marcel and show it to him. Once I am done with the
review operation as well. I will fix up a meeting with him next week to
plan the study.
Oh it feels nice to finally work on the prompts and everything else.
I also made sure that the choice in the choices step also shows up when
the user is using the arrow keys to go up or down.

I also added the runInAction thing to the controls. Since mobx seemed to
throwing a hissy fit when it the observable was being updated outside of
an action. Fixed the continuation prompt a bit as well.