Fixed css issues and focus issues
Fixed css issues for keycommandsmall so that the text of the key is
visble and readable. Also, fixed focus issues where focus is restored
back to the editor after selecting a choice. Made sure that the editor
is disabled when a prompt operation has started. I should be done with
rewrite soon. I can also add review as well. I still cant figure out a
way to focus the textbox on operation controls for some reason a call to
focus does not work.

Added a way to directly start operations using keyboard shortcuts by
using extensions.
Downgraded to tiptap@2.2.0 to fix the query selector issues in nodepos.
It seems to be working well and properly without any issues.
The editor looks disabled when an operation is loading.
Fixed the freeform form controls so that it relies completely on
instance controls.