Fixed focus issue on text controls
Fixed the focus issue on text control fields which was caused by
md-text-fields focus events not being processed properly changing the
element from md-text-field-filled to input. Made it so that the focus
events started working properly. Focusing the user exactly into the
input control.

Also, fixed the css for the input controls so that they are cohesive
similar to other elements of the prototype.