Fixed selection restore on operation cancel
The selection is now restored backt to where the user was when an
operation is being cancelled. This allows for more seamless interaction
for the DIY author.

Also, fixed the type error for starter-kit by removing @tiptap and
installing the packages back. This fixed the type issue since it seems
like it was probably leading to errors in pnpm lock. which is what lead
to incompatible types error.
But otherwise things seem to be on track i seem to be making steady
progress without major issues. The only issue being that needs to be
addressed is the history. but hopefully that will be resolved as well.
I need to focus on the rewrite operations.
Also, cleaned up the textcontrol from console.debug statements.