Updated event listeners
Updated the event listeners so that they are now fields that go to an
anonymous function for the event listener. This was done because a
regular method ends up losing the this context when hooked up to an
event listener.

Also, made it so that focus and blur psuedo-selection happens properly
its a bit finicky but overall it works fine.
Removed the event listener fields that I had made. In general cleaned
things up in text editor service.