Added Review operations and fixed minor bugs
Added review operations to review the entire diy tutorial or just a
portion of the DIY tutorial. Created a new review step and a review
operation specifically for these two operation. Also, created a review
step component and review tab component to allow the user access to the
reviews.

Chats and reviews are now saved along with the document so they are not
lost if the user quits in the middle there are still things that need to
be done there since restoring from hmr puts a constaint on how things
act.

Also, need to add logging and data collection. create a toolbar for the
users and also welcome dialogs and other popups. I can finally
concentrate on writing things as well for the thesis.

The file is now autosaved every 10 seconds.

I do need to improve how the reviews are being done since they seem to
rely a lot on previous context and seem to just respond to what was
going on previously in the conversation. I might also just try improving
the basic prompt instruction. The instruction for the selection is
pretty good and super specific. In general I can probably improve the
prompts.