Added dialogs service
Added dialog service mostly took code from wordcraft and rewrote the
welcome dialog and some of the other stuff as well. the api_server now
allows image uploads I need to create a dialog for it and test that we
can indeed insert images into the DIY tutorial.

But otherwise it seems things are fine otherwise.