Image uploads now work
We can now upload images onto the the editor. They work pretty well and
quickly. Though the only shame is that the LLM has no idea or knowledge
about these.

Had to work out some idiosyncracies and things for flask and image
upload with creating form data was pretty simple and good.
But otherwise finally it more or less seems complete.