Updated to GPT-4 and other minor fixes
Updated the chatgpt version to GPT 4. Since it seems to give a better
response in the review.

Also,the snackbar now appears when the editor is disabled and an
operation is currently pending and the user clicks the editor. Fixed the
tooltips and disabled the editor toolbar when an operation is pending.

Added the ability to finish the review operation since previous it was
being cancelled. I think I will move the saving of review to when an
operation is finished.

Updated some max-token lengths for certain prompts.