Added outline operation
Added outline operation. Fixed the image size issue in the editor page.
They are now restricted to a max-width of 600px. This was done so that
it does not take up too much space in writing the DIY tutorial. Fixed
the spellcheck and autocomplete for text area control. Fixed the id and
operation type for the FreeformStepControl.
Increased the snackbar timer to 10s