Adds PreTask and PostTask pages and data logging
Created pretask and post task pages. The pretask page has a demographic
form and the post task page allows me to export the participant data.

Also,I finally added logging to log all the datapoints. I am storing the
formdata,counters to count specific events. Though mostly the focus is
on operations. I added these everywhere. I created a new service just to
do that and it uses indexedDB instead of localStorage. I am using
indexedDB since it allows me to store the logRecords more easily and the
api is convenient. Though I am using a library called idb which makes it
more convenient to use the api in the form of async await.

I had to fix the ids for the operations. Since I really required that
for accessing what operation was being performed in a convenient manner.

I also fixed the reviewOperation and added ids and operationType to the
reviewOperations. Essentially the review is only saved to reviews if the
user clicks on enter to save.

Also, added a confirmation dialog if the user is trying to finish the
DIY tutorial.

I might still have to make some small fixes in terms of adjusting the
availability of operations. But I am ready for the study. I just need to
create a consent form and an instruction form.

I am finally done with this project. Now all I have to do is write
everything down, conduct the study and submit my thesis.