Added some guidance and help, fixed small bugs
Added some guidance and help and fixed small bugs like the error alert
in the outline page. Fixed the punctuation on the welcome dialog and
also made it so that the current DIY does not show up in the chat but
the user can toggle whether to send it or not using a toggle.

Also, added a INSERTED_WORD_COUNT which counts the amount of generated
words inserted.