Adds Image instruction operation
Adds the ability to generate a DIY instruction from an image input. Had
to modify text service and cursor service to figure out that the image
selection is a node selection. Added isNodeSelection and nodeSelection
as observables to contain whether the selection is a node selection or
not. Added a computed property which checked if it is a node image
selection. Added a new operation site for IMAGE_NODE_SELECTION.

Fixed the CSS for the image selection so that if the user has selected
the image there is dashed border around the image.

Created a new operation Image instruction operation. Also created a new
params type for this operation. Had to convert the image to a base64
string before sending the pre and post text as well to a chat completion
along with a prompt to craft a single DIY instruction. It seems to work
properly.

Had to update the GPT version from gpt-4-1106-preview to gpt-4-turbo so
that I could process images conveniently.