Adds Custom Image prompt operation
Adds a custom image prompt operation. This was fairly simple extension
of the image instruction prompt. Just added freeform controls to it.
Created a new operation type for it as well. Just overall removed the
debug stuff for production. Fixed the stop sequence issue for image
prompt with a custom stop sequence of <END> which results in a far nicer
output that is not cut short in the middle. I think with this I am now
confident with the prototype and dont have to do anything more.