Small fixes - Operation is available and logging
Fixed the Operation availability for generate intro and conclusion and
also review DIY operation. Added logging when a review is set.

This should be everything for the Study starting on monday. Now I just
have to write my thesis.

Select buttons are now disabled on submit.

Added a line about image prompt to the welcome dialog.

Created the final build for the prototype.