Final changes some css fixes for the large screen
Fix the chat tab where the input container wasnt showing up.
Fixed the css in the new diy page to better conform to CRAP so that
users do not mistake what the text field are about.
Fixed the text field on the demo page so that they are 100% width and
take the space of the entire form.