Fixed paths for backtracking
Fixed the url paths so that the flask server serves the frontend on any
route except for the predifined routes.
Added admin mode to the editor sidebar that allows going back to the
main page from the editor. Also, fixed the backtracking in the session
flow for the user. With this it should be easy to help the user go back
if they accidently confirmed the outline.